requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-dotenv
pytest
httpx==0.27.0
numpy
requests
